// web/src/components/dashboard/layout/types.ts
import { type LinkProps } from "react-router-dom";

type Team = {
    name: string;
    logo: React.ElementType;
    plan: string;
};

type BaseNavItem = {
    title: string;
    badge?: string;
    icon?: React.ElementType;
};

type NavLink = BaseNavItem & {
    url: LinkProps["to"] | (string & {});
    items?: never;
};

type NavCollapsible = BaseNavItem & {
    items: (BaseNavItem & { url: LinkProps["to"] | (string & {}) })[];
    url?: never;
};

type NavItem = NavCollapsible | NavLink;

type NavGroup = {
    title: string;
    items: NavItem[];
};

type SidebarData = {
    teams: Team[];
    navGroups: NavGroup[];
};

export type { SidebarData, NavGroup, NavItem, NavCollapsible, NavLink };